+++ app/api/members/delete-api/route.ts
@@ -4,11 +4,11 @@ import { NextResponse } from "next/server";
 
 export async function PATCH(
     req: Request,
-    { params }: { params: { memberId: string } }
 ) {
     try {
         const profile = await currentUserProfile();
         const { searchParams } = new URL(req.url);
+        const { memberId } = await req.json()
 
         const serverId = searchParams.get("serverId");
 
@@ -20,7 +20,7 @@ export async function PATCH(
         throw new Error("Server doesn't exist");
     }
 
-    if(!params.memberId) {
+    if(!memberId) {
         throw new Error("How did you even request this ?");
     }
 
@@ -32,7 +32,7 @@ export async function PATCH(
             Member: {
                 update: {
                     where: {
-                        id: params.memberId,
+                        id: memberId,
                         userProfileId: {
                         not: profile.id
                     },