+++ app/api/members/name-api/route.ts
@@ -4,14 +4,12 @@ import { NextResponse } from "next/server";
 
 export async function PATCH(
     req: Request,
-    { params }: { params: { memberId: string}}
 ) {
     try
     {
         const profile = await currentUserProfile();
         const { searchParams } = new URL(req.url);
-        const receivedForm = await req.json();
-        const  nickname2  = receivedForm.values.nickname;
+        const { memberId,nickname2 } = await req.json();
 
         // console.log("Nickname",receivedForm.values.nickname);
         const serverId2 = searchParams.get("serverId");
@@ -24,13 +22,13 @@ export async function PATCH(
             throw new Error("Server doesn't exist");
         }
 
-        if(!params.memberId) {
+        if(!memberId) {
             throw new Error("How did you even request this ?");
         }
 
         const member = await db.member.update({
             where: {
-                id:params.memberId,
+                id:memberId,
                 serverId:serverId2 as string,
                 userProfileId: profile.id
             },