+++ middleware.ts
@@ -20,6 +20,7 @@ export default authMiddleware({
     // Allow users visiting public routes to access them
     return NextResponse.next();
   },
+  // debug: true
 });
  
 export const config = {